Moving Accounts View to Settings View
