+++ App-iOS/Prueba de concepto/View/SettingsView.swift
@@ -7,6 +7,7 @@
 
 import SwiftUI
 import UserNotifications
+import WebKit
 
 // Vista para "Settings"
 struct SettingsView: View {
@@ -346,22 +347,95 @@ struct SuggestionFeedbackView: View {
 }
 
 
-struct UsefulLinksView: View {
-    @Binding var isPresented: Bool
-    var body: some View {
-        VStack {
-            Text("Useful Links")
-                .font(.title)
-                .fontWeight(.bold)
-                .padding()
+    struct WebView: UIViewRepresentable {
+        let urlString: String
 
-            // Add your useful links here
-            Link("Link 1", destination: URL(string: "https://www.link1.com")!)
-            Link("Link 2", destination: URL(string: "https://www.link2.com")!)
-            // Add more links as needed
-
-            Spacer()
+        func makeUIView(context: Context) -> WKWebView {
+            let webView = WKWebView()
+            return webView
         }
-        .padding()
+
+        func updateUIView(_ uiView: WKWebView, context: Context) {
+            if let url = URL(string: urlString) {
+                let request = URLRequest(url: url)
+                uiView.load(request)
+            }
+        }
     }
-}
+
+    struct UsefulLinksView: View {
+        @Binding var isPresented: Bool
+        @StateObject private var viewModel = AccountsViewModel()
+        @StateObject private var functions = GlobalFunctions()
+        @StateObject private var networkManager = NetworkMonitor()
+        @State private var isInternetConnected = true
+        
+        var body: some View {
+            ZStack() {
+                Color(red: 21/255, green: 191/255, blue: 129/255).edgesIgnoringSafeArea(.all)
+                VStack {
+                    Text("Useful links")
+                        .font(.title)
+                        .fontWeight(.bold)
+                        .foregroundColor(Color.white)
+                }
+            }.frame(maxWidth: 400, maxHeight: 60)
+            .background(Color(red: 240/255, green: 240/255, blue: 242/255))
+            VStack{
+                Text("Connect with")
+                    .fontWeight(.light)
+                    .foregroundColor(functions.isDaytime ? Color.black : Color.white)
+                List(viewModel.accounts, id: \.title) { account in
+                    Button(action: {
+                        if networkManager.isConnected {
+                            viewModel.selectedAccountURL = WebSheetItem(urlString: account.link)
+                        } else{
+                            viewModel.isAlertShowing = true
+                        }
+                    }) {
+                        AccountRow(account: account)
+                    }
+                }
+                .alert(isPresented: $viewModel.isAlertShowing) {
+                           Alert(
+                               title: Text("No Internet Connection"),
+                               message: Text("Please check your internet connection and try again."),
+                               dismissButton: .default(Text("OK"))
+                           )
+                       }
+            }.listStyle(PlainListStyle())
+                .background(functions.isDaytime ? Color.white : Color(red: 23/255, green: 24/255, blue: 25/255))
+                .sheet(item: $viewModel.selectedAccountURL) { webSheetItem in
+                NavigationView {
+                    WebView(urlString: webSheetItem.urlString)
+                        .navigationBarTitle("Account Login", displayMode: .inline)
+                        .navigationBarItems(leading: Button(action: {
+                            viewModel.selectedAccountURL = nil
+                        }) {
+                            Image(systemName: "arrow.left")
+                                .foregroundColor(.blue)
+                        })
+                }
+            }.onReceive(networkManager.$isConnected) { isConnected in
+                isInternetConnected = isConnected
+            }
+        }
+    }
+
+
+    struct AccountRow: View {
+        let account: Account
+        @StateObject private var functions = GlobalFunctions()
+        
+        var body: some View {
+            HStack {
+                Image(account.imageName)
+                    .resizable()
+                    .frame(width: 40, height: 40)
+                    .cornerRadius(20)
+                Text(account.title)
+                    .font(.headline)
+
+            }
+        }
+    }
